+++ planning/autoware_avoidance_target_detector/include/autoware/avoidance_target_detector/map_construction.hpp
@@ -12,8 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__AVOIDANCE_TARGET_DETECTOR__TRAFFIC_RULES_HPP_
-#define AUTOWARE__AVOIDANCE_TARGET_DETECTOR__TRAFFIC_RULES_HPP_
+#ifndef AUTOWARE__AVOIDANCE_TARGET_DETECTOR__MAP_CONSTRUCTION_HPP_
+#define AUTOWARE__AVOIDANCE_TARGET_DETECTOR__MAP_CONSTRUCTION_HPP_
+
+#include <autoware/route_handler/route_handler.hpp>
+
+#include <autoware_map_msgs/msg/lanelet_map_bin.hpp>
+#include <autoware_planning_msgs/msg/lanelet_route.hpp>
 
 #include <lanelet2_core/Forward.h>
 #include <lanelet2_routing/RoutingGraph.h>
@@ -23,7 +28,14 @@
 #include <optional>
 #include <string>
 
-namespace autoware::avoidance_target_detector::traffic_rules
+namespace autoware::avoidance_target_detector
+{
+
+using autoware::route_handler::RouteHandler;
+using autoware_map_msgs::msg::LaneletMapBin;
+using autoware_planning_msgs::msg::LaneletRoute;
+
+namespace traffic_rules
 {
 
 /** Location name registered for custom GoalPurposeRules traffic rules. */
@@ -62,6 +74,22 @@ std::optional<lanelet::ConstLanelet> get_left_lanelet(
 std::optional<lanelet::ConstLanelet> get_right_lanelet(
   const lanelet::routing::RoutingGraph & routing_graph, const lanelet::ConstLanelet & lanelet);
 
-}  // namespace autoware::avoidance_target_detector::traffic_rules
+}  // namespace traffic_rules
 
-#endif  // AUTOWARE__AVOIDANCE_TARGET_DETECTOR__TRAFFIC_RULES_HPP_
+/**
+ * @brief Build the route handler and goal-purpose routing graph from a map and route.
+ * @details Creates the route handler internally, sets the map and route, and builds the cached
+ *          routing graph. Outputs are written into the provided references.
+ * @param map_bin Lanelet map binary message.
+ * @param route Lanelet route message.
+ * @param route_handler Output route handler (created internally).
+ * @param routing_graph Output goal-purpose routing graph.
+ */
+void create_map(
+  const LaneletMapBin & map_bin, const LaneletRoute & route,
+  std::shared_ptr<RouteHandler> & route_handler,
+  lanelet::routing::RoutingGraphConstPtr & routing_graph);
+
+}  // namespace autoware::avoidance_target_detector
+
+#endif  // AUTOWARE__AVOIDANCE_TARGET_DETECTOR__MAP_CONSTRUCTION_HPP_
